Page create site page

+++ database/migrations/2025_05_03_165459_create_role_limits_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->id();
             $table->unsignedBigInteger('role_id');
             $table->string('name');
-            $table->integer('value');
+            $table->unsignedBigInteger('value');
             $table->timestamps();
 
             $table->foreign('role_id')->references('id')->on('roles')->onDelete('cascade');
